requirements.txt
Django==3.1.3
djangorestframework==3.12.2
psycopg2-binary==2.9.1 
django-db-multitenant==0.3.2 
django-cors-headers==3.5.0
